requirements.txt
fastapi
uvicorn
google-cloud-storage
psycopg2-binary
